requirements.txt
absl-py==0.13.0
astor==0.8.1
cached-property==1.5.2
click==8.0.1
colorama==0.4.4
Flask==2.0.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.39.0
gunicorn==20.1.0
h5py==2.10.0
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.21.1
opt-einsum==3.3.0
Pillow==8.3.1
protobuf==3.17.3
PyYAML==5.4.1
scipy==1.7.0
six==1.16.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
